+++ packages/client/hooks/useTipTapTaskEditor.ts
@@ -1,5 +1,7 @@
+import Highlight from '@tiptap/extension-highlight'
 import {TaskItem, TaskList} from '@tiptap/extension-list'
 import Mention from '@tiptap/extension-mention'
+import {TextStyleKit} from '@tiptap/extension-text-style'
 import {Placeholder} from '@tiptap/extensions'
 import {Extension, generateText, useEditor} from '@tiptap/react'
 import StarterKit from '@tiptap/starter-kit'
@@ -39,6 +41,8 @@ export const useTipTapTaskEditor = (
       content: contentJSON,
       extensions: [
         StarterKit.configure({link: false}),
+        TextStyleKit,
+        Highlight,
         FileUpload.configure({
           scopeKey: teamId,
           assetScope: 'Team',