+++ packages/server/graphql/public/mutations/addIntegrationProvider.ts
@@ -65,6 +65,13 @@ const addIntegrationProvider: MutationResolvers['addIntegrationProvider'] = asyn
       }
     }
   }
+  if (context.resourceGrants) {
+    if (scope === 'org' && !(await context.resourceGrants.hasOrg(orgId!))) {
+      return {error: {message: 'PAT does not grant access to this organization'}}
+    } else if (scope === 'team' && !(await context.resourceGrants.hasTeam(teamId!))) {
+      return {error: {message: 'PAT does not grant access to this team'}}
+    }
+  }
 
   // VALIDATION
   const {