+++ packages/server/graphql/public/types/UpdatePersonalAccessTokenSuccess.ts
@@ -1,0 +1,13 @@
+import type {UpdatePersonalAccessTokenSuccessResolvers} from '../resolverTypes'
+
+export type UpdatePersonalAccessTokenSuccessSource = {
+  patId: string
+}
+
+const UpdatePersonalAccessTokenSuccess: UpdatePersonalAccessTokenSuccessResolvers = {
+  personalAccessToken: async ({patId}, _args, {dataLoader}) => {
+    return dataLoader.get('personalAccessTokens').loadNonNull(patId)
+  }
+}
+
+export default UpdatePersonalAccessTokenSuccess